Remove `on: create` from CI
This doesn't appear to actually take a `tags` sub-argument, it's just
on-create of either a tag or a branch. This means that contributors
pushing to their own repositories with feature branches trigger builds,
but only on branch creation and not on further commits. This switches to
`on: push: tags` which should retain the behavior of running on tags.